+Removed test for account created callback for the moment +Renamed abstract-adapter to abstract
