+++ src/adapters/abstract.js
