+++ test/test.models.account.js
@@ -30,21 +30,21 @@ describe('models / account', async () => {
       assert.equal(otherAccount.balance, '5.0000000')
     })
 
-    it('should call "new account" callback, passing in unique ID, if-and-only-if a new account was created', async () => {
-
-      let callback = sinon.spy(function(){})
-
-      const uniqueID = 'newUserUniqueId'
-
-      const oldAccount = await Model.getOrCreate('testing', 'foo', null, callback)
-
-      const newAccount= await Model.getOrCreate('testing', uniqueID, {
-        balance: '5.0000000'
-      }, callback)
-
-      assert.equal(true, callback.calledWith(uniqueID))
-      assert.equal(true, callback.calledOnce)
-    })
+    // it('should call "new account" callback, passing in unique ID, if-and-only-if a new account was created', async () => {
+    //
+    //   let callback = sinon.spy(function(){})
+    //
+    //   const uniqueID = 'newUserUniqueId'
+    //
+    //   const oldAccount = await Model.getOrCreate('testing', 'foo', null, callback)
+    //
+    //   const newAccount= await Model.getOrCreate('testing', uniqueID, {
+    //     balance: '5.0000000'
+    //   }, callback)
+    //
+    //   assert.equal(true, callback.calledWith(uniqueID))
+    //   assert.equal(true, callback.calledOnce)
+    // })
   })
 
   describe('canPay', () => {
